// src/core/api/generated/spring/schemas/CallImportJobDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */
import type { CallImportJobDTOStatus } from './CallImportJobDTOStatus';
import type { ImportResult } from './ImportResult';

export interface CallImportJobDTO {
  jobId?: string;
  filename?: string;
  status?: CallImportJobDTOStatus;
  stage?: string;
  totalRows?: number;
  processedRows?: number;
  successCount?: number;
  duplicateCount?: number;
  failedCount?: number;
  validationErrorCount?: number;
  masterMissingCount?: number;
  systemErrorCount?: number;
  errorMessage?: string;
  createdBy?: string;
  createdDate?: string;
  completedDate?: string;
  progressPercentage?: number;
  result?: ImportResult;
}